requirements.txt
redis==5.0.1
psutil==5.9.8
GitPython==3.1.44
requests==2.31.0